Use NSInteger and NSUInteger.

+++ KeywordMapping.m
@@ -1,3 +1,4 @@
+#import "KeywurlPlugin.h"
 #import "KeywordMapping.h"
 
 @implementation QueryToken
@@ -100,8 +101,8 @@
     int searchStart = 0;
     while (start < [expansion length]) {
         NSRange searchRange = NSMakeRange(searchStart, [expansion length] - searchStart);
-        int curlStart = [expansion rangeOfString: @"{" options: 0 range: searchRange].location;
-        int curlEnd = curlStart == NSNotFound ? NSNotFound : [expansion rangeOfString: @"}" options: 0 
+        NSInteger curlStart = [expansion rangeOfString: @"{" options: 0 range: searchRange].location;
+        NSInteger curlEnd = curlStart == NSNotFound ? NSNotFound : [expansion rangeOfString: @"}" options: 0 
             range: NSMakeRange(curlStart, [expansion length] - curlStart)].location;
         if (curlEnd != NSNotFound) {
             NSString* symbol = [expansion substringWithRange: NSMakeRange(curlStart + 1, curlEnd - curlStart - 1)];
